Make tooltip interactive.
The tooltip can display so much data that it is impossible to take it
all in before it goes away when hovering over the applet.  This change
makes the tooltip interactive so that the user can actually read the
data before is vanishes.

This change also restores the display of cached memory in the memory bar
chart.  These two changes restore the behavior of this applet before
introducing Plasma required this new one.

Be sure to configure General->Monitor type->Compact bar and Colors->Set
Colors Manually.  The defaults are much less appealing.